requirements.txt
django==1.5.1
django-admin-bootstrapped
south==0.8.1
dj-database-url
psycopg2
djangorestframework==2.3.5